other/requirements.txt
discord>=1.7.3
python-kucoin>=2.1.2
